+++ arviz/plots/backends/bokeh/densityplot.py
@@ -12,8 +12,7 @@ from ...plot_utils import (
     calculate_point_estimate,
 )
 from ....stats import hpd
-from ....stats.stats_utils import histogram, get_bins
-from ....kde_utils import _fast_kde
+from ....numeric_utils import _fast_kde, histogram, get_bins
 
 
 def plot_density(
