+++ arviz/plots/backends/bokeh/ppcplot.py
@@ -5,7 +5,7 @@ from . import backend_kwarg_defaults
 from .. import show_layout
 from ...kdeplot import plot_kde, _fast_kde
 from ...plot_utils import _create_axes_grid
-from ....stats.stats_utils import histogram, get_bins
+from ....numeric_utils import histogram, get_bins
 
 
 def plot_ppc(
