+++ arviz/plots/backends/matplotlib/densityplot.py
@@ -9,8 +9,7 @@ from ...plot_utils import (
     _create_axes_grid,
     calculate_point_estimate,
 )
-from ....stats.stats_utils import get_bins
-from ....kde_utils import _fast_kde
+from ....numeric_utils import _fast_kde, get_bins
 
 
 def plot_density(
