+++ arviz/plots/backends/matplotlib/ppcplot.py
@@ -9,8 +9,7 @@ from ...plot_utils import (
     make_label,
     _create_axes_grid,
 )
-from ....stats.stats_utils import histogram, get_bins
-from ....kde_utils import _fast_kde
+from ....numeric_utils import _fast_kde, histogram, get_bins
 
 
 def plot_ppc(
