+++ arviz/plots/backends/matplotlib/traceplot.py
@@ -10,7 +10,7 @@ from . import backend_kwarg_defaults, backend_show
 from ...distplot import plot_dist
 from ...rankplot import plot_rank
 from ...plot_utils import _scale_fig_size, make_label, format_coords_as_labels
-from ....stats.stats_utils import get_bins
+from ....numeric_utils import get_bins
 
 
 def plot_trace(
